# Optimize `view_engineer` to use DB-side aggregation and cut load time