… dojo/metrics/views.py