> Initially inspired by [#12575](https://github.com/DefectDojo/django-DefectDojo/issues/12575)  

---

#### ✨  Key Points
* **Full rewrite of `view_engineer`**
  * replaces Python loops with SQL `COUNT` / `CASE` / `Subquery`
  * helper functions: `_age_buckets`, `_product_stats`, `_augment_series_with_accepted`
  * preserves the 5-minute `@cache_page` (But now it **can be removed** and I guess - should be removed)
* **One SQL round-trip per metric**
  * month/week open / closed / accepted
  * product severity tables
  * age buckets for high/critical
* **Targeted ORM tuning**
  * lean `select_related` & `prefetch_related`
  * dead imports & paths removed
* **No template or migration changes**

---

#### 🚀  Performance on our dataset
**I cannot run it on the our production dataset for now, but I excpect huge improvement**

| Env            | Before | After | Δ |
|----------------|--------|-------|---|
| dev | **12.7s** | **32 ms** | –99.7 % |
| prod | **≈ 190s** | - | - |

<details>
<summary>Screenshots</summary>

| Before | After |
|:--:|:--:|
| <img width="479" alt="Screenshot 2025-06-14 at 23 31 00" src="https://github.com/user-attachments/assets/b1a8323d-3216-420d-b761-1b3c598069a2" /> | <img width="481" alt="Screenshot 2025-06-14 at 23 30 29" src="https://github.com/user-attachments/assets/9496e552-bf9a-467d-8d67-3583fd59c7fc" />|

</details>