@valentijnscholten 2025-06-17
Thanks again for a nice optimization PR. If the queries are now much faster, the caching can probably be removed.

Since these changes are a little more involved. Is there anyway we can verify the results of the queries are still the same/correct. Would it be feasible to add a little unit test similar to https://github.com/DefectDojo/django-DefectDojo/blob/bugfix/unittests/test_metrics_queries.py ?