@valentijnscholten 2025-06-17
Something I also wondered if there is a better/clearer name to the `stuff` named variables. To make it more clear what they are.